Move validatePack to pack package
